+++ data/freight/gqgv/cleaned.py
@@ -8,6 +8,7 @@ RENAMES = {"ORIGIN":"origin_nuts_id",
            "COUNTRY_OF_UNLOADING": "destination_country",
            "VEHICLE_TYPE": "vehicle_type",
            "TYPE_OF_GOOD": "good_type",
+           "KM_PERFORMANCE": "distance_km",
            "WEIGHTING_FACTOR": "weight",
            "DIVISOR": "divisor"
            }
@@ -15,7 +16,7 @@ RENAMES = {"ORIGIN":"origin_nuts_id",
 FIELDS = ["origin_nuts_id","destination_nuts_id",
           "origin_municipality", "destination_municipality",
           "origin_country", "destination_country",
-          "vehicle_type", "good_type", "weight"
+          "vehicle_type", "good_type", "distance_km", "weight"
           ]
 
 # VEHICLE_TYPES = {
@@ -39,7 +40,7 @@ def configure(context, require):
 def execute(context):
     df = context.stage("data.freight.gqgv.raw")
 
-    # rename
+    # rename columns
     df = df.rename(RENAMES, axis=1)
 
     # apply divisor to weight
@@ -50,13 +51,13 @@ def execute(context):
 
     # There are some NUTS ids that do not exist in our NUTS data (maybe old ids)
     # for now, drop all trips where NUTS not in NUTS data
-    print("Dropping all trips where NUTS id not contained in NUTS data ...")
+    print("Dropping all OD pairs where NUTS id not contained in NUTS data ...")
     number_trips = len(df)
     df_nuts = context.stage("data.spatial.nuts")
     nuts_ids = list(df_nuts["nuts_id"].unique())
     df = df[(df["origin_nuts_id"].isin(nuts_ids)) & (df["destination_nuts_id"].isin(nuts_ids))]
     number_trips_dropped = number_trips - len(df)
-    print("Dropped %s of %s trips" % (number_trips_dropped, number_trips))
+    print("Dropped %s of %s OD pairs" % (number_trips_dropped, number_trips))
 
     # package
     df = df[FIELDS]