+++ data/freight/gqgv/od.py
@@ -4,6 +4,8 @@ import numpy as np
 
 def configure(context, require):
     require.stage("data.freight.gqgv.cleaned")
+    require.stage("data.freight.scaling_factor")
+    require.config("enable_scaling")
 
 
 def execute(context):
@@ -28,6 +30,10 @@ def execute(context):
 
         # compute demand
         demands[vehicle_type] = int(np.round(np.sum(matrix_values) / number_of_days))
+
+        # scale demand
+        if context.config["enable_scaling"]:
+            demands[vehicle_type] *= context.stage("data.freight.scaling_factor")
 
         # make sure each from sums up to one
         f_zero = np.sum(matrix_values, axis=1) == 0.0